abstract item model building into a single copy of the meshing code
