+++ fabric/src/main/java/net/caffeinemc/mods/sodium/fabric/render/ImprovedItemModelBuilder.java
@@ -7,130 +7,12 @@ import net.minecraft.client.resources.model.ModelBaker;
 import net.minecraft.client.resources.model.cuboid.CuboidFace;
 import net.minecraft.client.resources.model.cuboid.FaceBakery;
 import net.minecraft.client.resources.model.geometry.BakedQuad;
-import net.minecraft.client.resources.model.geometry.QuadCollection;
-import org.joml.Vector3f;
+import net.minecraft.core.Direction;
+import org.joml.Vector3fc;
 
-import static net.minecraft.client.resources.model.cuboid.ItemModelGenerator.*;
-
-public class ImprovedItemModelBuilder {
-
-    public static void bakeSideQuads(
-            QuadCollection.Builder builder,
-            ModelBaker.Interner interner,
-			BakedQuad.MaterialInfo materialInfo,
-			ModelState modelState
-	) {
-		var sprite = materialInfo.sprite().contents();
-
-		var xScale = 16.0F / sprite.width();
-		var yScale = 16.0F / sprite.height();
-
-		var layerBias = materialInfo.tintIndex() * ImprovedItemModelBuilderBase.LAYER_XY_BIAS;
-
-		for (ImprovedItemModelBuilderBase.SideFace sideFace : ImprovedItemModelBuilderBase.buildSideFaces(sprite)) {
-			var faceFacing = sideFace.facing();
-			var faceAnchor = sideFace.anchor();
-			var faceMin = sideFace.min();
-			var faceMax = sideFace.max();
-
-            // Calculate the start coordinate and length of the side quad using the side face properties, as described
-            // in the diagram in FaceStorage.
-			float minX = faceFacing.isHorizontal() ? faceMin : faceAnchor;
-			float minY = faceFacing.isHorizontal() ? faceAnchor : faceMin;
-			float length = faceMax - faceMin + 1.0F;
-
-			var u0 = 0.0F;
-			var v0 = 0.0F;
-
-			var u1 = 0.0F;
-			var v1 = 0.0F;
-
-			if (faceFacing.isHorizontal()) {
-				u0 = minX + UV_SHRINK;
-				v0 = minY + UV_SHRINK;
-				u1 = minX + length - UV_SHRINK;
-				v1 = minY + 1.0F - UV_SHRINK;
-			} else {
-				u0 = minX + UV_SHRINK;
-				v0 = minY + length - UV_SHRINK;
-				u1 = minX + 1.0F - UV_SHRINK;
-				v1 = minY + UV_SHRINK;
-			}
-
-			var fromX = minX;
-			var fromY = minY;
-			var toX = minX;
-			var toY = minY;
-
-			switch (faceFacing) {
-				case UP -> {
-                    toX = minX + length;
-                }
-				case LEFT -> {
-                    toY = minY + length;
-                }
-				case DOWN -> {
-					fromY = minY + 1.0F;
-					toY = minY + 1.0F;
-					toX = minX + length;
-				}
-				case RIGHT -> {
-					fromX = minX + 1.0F;
-					toX = minX + 1.0F;
-					toY = minY + length;
-				}
-			}
-
-			fromX *= xScale;
-			fromY *= yScale;
-			toX *= xScale;
-			toY *= yScale;
-
-			fromY = 16.0F - fromY;
-			toY = 16.0F - toY;
-
-            // Expand side quads along their normals to cover the layer below without gaps
-            switch (faceFacing) {
-                case UP -> {
-                    fromX -= layerBias;
-                    toX += layerBias;
-                    fromY = toY = fromY + layerBias;
-                }
-                case DOWN -> {
-                    fromX -= layerBias;
-                    toX += layerBias;
-                    fromY = toY = toY - layerBias;
-                }
-                case LEFT -> {
-                    fromY += layerBias;
-                    toY -= layerBias;
-                    fromX = toX = fromX - layerBias;
-                }
-                case RIGHT -> {
-                    fromY += layerBias;
-                    toY -= layerBias;
-                    fromX = toX = toX + layerBias;
-                }
-            }
-
-			builder.addUnculledFace(
-					FaceBakery.bakeQuad(
-							interner,
-							new Vector3f(fromX, fromY, MIN_Z),
-							new Vector3f(toX, toY, MAX_Z),
-							new CuboidFace.UVs(
-									u0 * xScale,
-									v0 * yScale,
-									u1 * xScale,
-									v1 * yScale
-							),
-							Quadrant.R0,
-							materialInfo,
-							faceFacing.getDirection(),
-							modelState,
-							null
-					)
-			);
-		}
-	}
+public class ImprovedItemModelBuilder extends ImprovedItemModelBuilderBase<Void> {
+    @Override
+    public BakedQuad bakeQuad(ModelBaker.Interner interner, Vector3fc from, Vector3fc to, CuboidFace.UVs uvs, BakedQuad.MaterialInfo materialInfo, Direction normal, ModelState modelState, Void extraFaceData) {
+        return FaceBakery.bakeQuad(interner, from, to, uvs, Quadrant.R0, materialInfo, normal, modelState, null);
+    }
 }
