+++ fabric/src/main/java/net/caffeinemc/mods/sodium/mixin/features/render/model/ItemModelGeneratorMixin.java
@@ -9,12 +9,15 @@ import net.minecraft.client.resources.model.cuboid.ItemModelGenerator;
 import net.minecraft.client.resources.model.geometry.BakedQuad;
 import net.minecraft.client.resources.model.geometry.QuadCollection;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 
 @Mixin(ItemModelGenerator.class)
 public class ItemModelGeneratorMixin {
+    @Unique
+    private static final ImprovedItemModelBuilder sodium$INSTANCE = new ImprovedItemModelBuilder();
 
     @WrapMethod(method = "bakeSideFaces")
     private static void improvedBakeSideFaces(QuadCollection.Builder builder, ModelBaker.Interner interner, ModelState modelState, BakedQuad.MaterialInfo materialInfo, Operation<Void> original) {
-        ImprovedItemModelBuilder.bakeSideQuads(builder, interner, materialInfo, modelState);
+        sodium$INSTANCE.bakeSideQuads(builder, interner, materialInfo, modelState, null);
     }
 }
